requirements.txt
aiogram>=3.13,<4
aiosqlite>=0.20.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
